make generic burn_state_t constructor type-bound
This has the effect of preventing type extension because type
extension facilitates inheritance nad inheritance is a somewhat
deprecated feature in the object-oriented design community
(cf. Gamma et al., 1994).  This will be a project-wide decision
that can easily be reversed, if so desired, by making the
generic constructor non-type-bound. One exception: we will reserve
type extension for future use in extending abstract types.